requirements.txt
Flask==2.0.2
psycopg2-binary==2.9.8
requests==2.27.1
Werkzeug==2.2.2
toml




